Consolidate solvers and update examples
